my profiles update

+++ src/controllers/videoController.js
@@ -1,31 +1,48 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
-    const videos = await Video.find({}).sort({ createdAt: "desc" });
-    return res.render("home", { pageTitle: "Home", videos });
+    const videos = await Video.find({}).sort({createdAt: "desc"});
+    return res.render("home", {pageTitle: "Home", videos});
 };
+
 export const watch = async (req, res) => {
     const {id} = req.params;
-    const video = await Video.findById(id);
-    if(video) {
-        return res.render("watch", { pageTitle: video.title, video });
+    const video = await Video.findById(id).populate("owner");
+    console.log(video);
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
     }
-    return res.status(404).render("404", { pageTitle: "Video not found." });
+    return res.render("watch", {pageTitle: video.title, video});
 };
+
 export const getEdit = async (req, res) => {
     const {id} = req.params;
+    const {
+        user: {_id},
+    } = req.session;
     const video = await Video.findById(id);
-    if(!video) {
-        return res.status(404).render("404", { pageTitle: "Video not found." });
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
     }
-    return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+    return res.render("edit", {pageTitle: `Edit ${video.title}`, video});
 };
+
 export const postEdit = async (req, res) => {
     const {id} = req.params;
+    const {
+        user: {_id},
+    } = req.session;
     const {title, description, hashtags} = req.body;
-    const video = await Video.exists({ _id: id });
-    if(!video) {
-        return res.status(404).render("404", { pageTitle: "Video not found." });
+    const video = await Video.exists({_id: id});
+    if (!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -34,34 +51,58 @@ export const postEdit = async (req, res) => {
     });
     return res.redirect(`/videos/${id}`);
 };
+
 export const getUpload = (req, res) => {
     return res.render("upload", {pageTitle: "Upload Video"});
 };
+
 export const postUpload = async (req, res) => {
-    const file = req.file;
+    const {
+        user: {_id},
+    } = req.session;
+    const {path: fileUrl} = req.file;
     const {title, description, hashtags} = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
-            fileUrl: file.path,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags)
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
-    } catch(error) {
+    } catch (error) {
         return res.status(400).render("upload", {
             pageTitle: "Upload Video",
             errorMessage: error._message
         });
     }
 };
+
 export const deleteVideo = async (req, res) => {
-    const { id } = req.params;
+    const {id} = req.params;
+    const {
+        user: {_id},
+    } = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if(!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."});
+    }
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(id), 1);
+    user.save();
     return res.redirect("/");
 }
+
 export const search = async (req, res) => {
-    const { keyword } = req.query;
+    const {keyword} = req.query;
     let videos = [];
     if (keyword) {
         videos = await Video.find({
@@ -70,5 +111,5 @@ export const search = async (req, res) => {
             },
         });
     }
-    return res.render("search", { pageTitle: "Search", videos });
+    return res.render("search", {pageTitle: "Search", videos});
 }
